# frontend/travel-with-anna-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification for Travel with Anna API",
    "description": "OpenApi documentation for Travel with Anna API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/twa/v1",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/trip": {
      "get": {
        "tags": [
          "Trip"
        ],
        "operationId": "getAllOwnersTrips",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseTripResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trip"
        ],
        "operationId": "createTrip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreatorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trip"
        ],
        "operationId": "deleteTrip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Trip"
        ],
        "operationId": "updateTrip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/expanse": {
      "post": {
        "tags": [
          "Expanse"
        ],
        "operationId": "createOrUpdateExpanse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpanseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpanseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/expanse/trip-currency-values": {
      "post": {
        "tags": [
          "Expanse"
        ],
        "operationId": "getTripCurrencyValues",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCurrencyValuesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpanseInTripCurrency"
                }
              }
            }
          }
        }
      }
    },
...
    "/user": {
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/change-password": {
      "patch": {
        "tags": [
          "User"
        ],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/trip/{tripId}": {
      "get": {
        "tags": [
          "Trip"
        ],
        "operationId": "getTripById",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/pdf/reports/trip/{tripId}": {
      "get": {
        "tags": [
          "Pdf Report"
        ],
        "operationId": "generateTripPdfReport",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
    "/pdf/reports/expanse/{tripId}": {
      "get": {
        "tags": [
          "Pdf Report"
        ],
        "operationId": "generateExpansePdfReport",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/expanse/{expanseId}": {
      "get": {
        "tags": [
          "Expanse"
        ],
        "operationId": "getExpanseById",
        "parameters": [
          {
            "name": "expanseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpanseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/expanse/{entityId}/type/{entityType}": {
      "get": {
        "tags": [
          "Expanse"
        ],
        "operationId": "getExpanseByEntity",
        "parameters": [
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "entityType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpanseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/expanse/exchange": {
      "get": {
        "tags": [
          "Expanse"
        ],
        "operationId": "getExchangeRate",
        "parameters": [
          {
            "name": "currencyFrom",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currencyTo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/day/trip/{tripId}": {
      "get": {
        "tags": [
          "Day"
        ],
        "operationId": "getDays",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/budget/{budgetId}/expanses/{tripId}": {
      "get": {
        "tags": [
          "Budget"
        ],
        "operationId": "getBudgetExpanses",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "budgetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetExpensesRespond"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TripCreatorRequest": {
        "required": [
          "currency",
          "toSpend"
        ],
        "type": "object",
        "properties": {
          "tripName": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string"
          },
          "currency": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          },
          "toSpend": {
            "type": "number"
          }
        }
      },
...
      "ExpanseRequest": {
        "required": [
          "currency",
          "exchangeRate",
          "expanseCategory",
          "paid",
          "price"
        ],
        "type": "object",
        "properties": {
          "expanseId": {
            "type": "integer",
            "format": "int64"
          },
          "entityId": {
            "type": "integer",
            "format": "int64"
          },
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "expanseName": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string"
          },
          "expanseCategory": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "currency": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          },
          "price": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "paid": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "exchangeRate": {
            "minimum": 0.0,
            "exclusiveMinimum": true,
            "type": "number"
          },
          "priceInTripCurrency": {
            "type": "number"
          },
          "paidInTripCurrency": {
            "type": "number"
          },
          "entityType": {
            "type": "string"
          }
        }
      },
      "ExpanseResponse": {
        "type": "object",
        "properties": {
          "expanseId": {
            "type": "integer",
            "format": "int64"
          },
          "expanseName": {
            "type": "string"
          },
          "expanseCategory": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "paid": {
            "type": "number"
          },
          "exchangeRate": {
            "type": "number"
          },
          "priceInTripCurrency": {
            "type": "number"
          },
          "paidInTripCurrency": {
            "type": "number"
          }
        }
      },
      "TripCurrencyValuesRequest": {
        "required": [
          "exchangeRate",
          "paid",
          "price"
        ],
        "type": "object",
        "properties": {
          "price": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "paid": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "exchangeRate": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number"
          }
        }
      },
      "ExpanseInTripCurrency": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number"
          },
          "paid": {
            "type": "number"
          }
        }
      },
      "DayGeneratorRequest": {
        "required": [
          "endDate",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "associatedTripId": {
            "type": "integer",
            "format": "int64"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        }
      },
...
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "UserCredentialsRequest": {
        "required": [
          "email",
          "userName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "userName": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "ChangePasswordRequest": {
        "required": [
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "TripEditRequest": {
        "type": "object",
        "properties": {
          "dayGeneratorRequest": {
            "$ref": "#/components/schemas/DayGeneratorRequest"
          },
          "tripName": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "PageResponseTripResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "TripResponse": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "backpackId": {
            "type": "integer",
            "format": "int64"
          },
          "budgetId": {
            "type": "integer",
            "format": "int64"
          },
          "tripName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "amountOfDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "noteId": {
            "type": "integer",
            "format": "int64"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "ExchangeResponse": {
        "type": "object",
        "properties": {
          "errorMsg": {
            "type": "string"
          },
          "exchangeRate": {
            "type": "number"
          }
        }
      },
      "ActivityListResponse": {
        "type": "object",
        "properties": {
          "activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityResponse"
            }
          }
        }
      },
      "ActivityResponse": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int64"
          },
          "activityTitle": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "badge": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "note": {
            "$ref": "#/components/schemas/NoteResponse"
          },
          "expanse": {
            "$ref": "#/components/schemas/ExpanseResponse"
          },
          "associatedId": {
            "type": "integer",
            "format": "int64"
          },
          "address": {
            "$ref": "#/components/schemas/AddressResponse"
          },
          "dayTag": {
            "type": "boolean"
          }
        }
      },
      "AddressResponse": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "integer",
            "format": "int64"
          },
          "place": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "DayResponse": {
        "type": "object",
        "properties": {
          "dayId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "isToday": {
            "type": "boolean"
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32"
          },
          "note": {
            "$ref": "#/components/schemas/NoteResponse"
          },
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "activities": {
            "$ref": "#/components/schemas/ActivityListResponse"
          }
        }
      },
      "BudgetResponse": {
        "type": "object",
        "properties": {
          "budgetId": {
            "type": "integer",
            "format": "int64"
          },
          "currency": {
            "type": "string"
          },
          "budgetAmount": {
            "type": "number"
          },
          "tripId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BudgetExpensesRespond": {
        "type": "object",
        "properties": {
          "budgetResponse": {
            "$ref": "#/components/schemas/BudgetResponse"
          },
          "expanses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpanseResponse"
            }
          },
          "sumsByCurrency": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpanseByCurrency"
            }
          },
          "overallPriceInTripCurrency": {
            "type": "number"
          },
          "overallPaidInTripCurrency": {
            "type": "number"
          },
          "totalDebtInTripCurrency": {
            "type": "number"
          },
          "priceBalanceDue": {
            "type": "number"
          },
          "paidBalanceDue": {
            "type": "number"
          }
        }
      },
      "ExpanseByCurrency": {
        "type": "object",
        "properties": {
          "currencyCode": {
            "type": "string"
          },
          "totalPrice": {
            "type": "number"
          },
          "totalPaid": {
            "type": "number"
          },
          "totalPriceInTripCurrency": {
            "type": "number"
          },
          "totalPaidInTripCurrency": {
            "type": "number"
          },
          "totalDebt": {
            "type": "number"
          }
        }
      },
...
      "PasswordRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "TripRequest": {
        "required": [
          "tripId"
        ],
        "type": "object",
        "properties": {
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}